% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{activityInfoAuthentication}
\alias{activityInfoAuthentication}
\title{Constructs a httr::authentication object from saved credentials 
from the user's home directory at ~/.activityinfo.credentials}
\usage{
activityInfoAuthentication(newValue)
}
\description{
Constructs a httr::authentication object from saved credentials 
from the user's home directory at ~/.activityinfo.credentials
}
