% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{activityInfoLogin}
\alias{activityInfoLogin}
\title{Authenticate and store user credentials to authorize requests for the
ActivityInfo API}
\usage{
activityInfoLogin(userEmail, password)
}
\arguments{
\item{userEmail}{The email address used to log in with to ActivityInfo.org}

\item{password}{The user's ActivityInfo password}
}
\description{
Configures the current session to use a user's email address and
password for authentication to \href{www.activityinfo.org}{ActivityInfo.org}
}
\examples{
\dontrun{
activityInfoLogin("nouser@activityinfo.org", "pass")
}
}
