% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{activityInfoRootUrl}
\alias{activityInfoRootUrl}
\title{Get or set the root url for the ActivityInfo}
\usage{
activityInfoRootUrl(newUrl, new.url)
}
\arguments{
\item{newUrl}{The new URL to set as the ActivityInfo root URL}

\item{new.url}{Deprecated: please use newUrl. The new URL to set as the 
ActivityInfo root URL}
}
\description{
This call gets or sets the root url used for a session, \emph{valid only
during the session}.
}
