% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{activityInfoToken}
\alias{activityInfoToken}
\title{Store personal token to authorize requests for the
ActivityInfo API}
\usage{
activityInfoToken(token)
}
\arguments{
\item{token}{The personal token used to authenticate with to ActivityInfo.org}
}
\description{
Configures the current session to use a personal token for authentication to \href{www.activityinfo.org}{ActivityInfo.org}
}
\examples{
\dontrun{
activityInfoToken("<API TOKEN>")
}
}
