% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{activityInfoVariableExpression}
\alias{activityInfoVariableExpression}
\title{An ActivityInfo variable expression}
\usage{
activityInfoVariableExpression(x, path = NULL)
}
\arguments{
\item{x}{An ActivityInfo form id, form tree, or another ActivityInfo variable expression.}

\item{path}{The next valid element/name in the variable expression or the root form ID (optional).}
}
\description{
A helper object to get a reference to variables/columns available from a form
or reference field. These expressions support the $ syntax in R so that the
user can interactively expand reference and sub-form fields.

This is also returned when using $ in a getRecords() object.
}
