% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{addDatabase}
\alias{addDatabase}
\title{addDatabase}
\usage{
addDatabase(label, databaseId = cuid())
}
\arguments{
\item{label}{The new database label}

\item{databaseId}{The new database identifier; a cuid will be generated if missing}
}
\description{
Adds a new database.
}
\examples{
\dontrun{
newDb <- addDatabase("Programme information system")
}
}
