% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{addFilter}
\alias{addFilter}
\title{Add an ActivityInfo API style filter}
\usage{
addFilter(x, formulaFilter)
}
\arguments{
\item{x}{the remote records object fetched with getRecords().}

\item{formulaFilter}{a character string with the formula.}
}
\description{
This adds a filter to an ActivityInfo remote records object. Usually one
would use the dplyr::filter() verb instead. The dplyr::filter() verb attempts
to translate R expressions into ActivityInfo style filters. But if this is
not successful, it can be useful to add this filter formula using addFilter()
}
