% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{adjustWindow}
\alias{adjustWindow}
\title{Adjust the offset and limit of a remote records object}
\usage{
adjustWindow(x, offSet = 0L, limit)
}
\arguments{
\item{x}{the remote records object fetched with \link{getRecords}.}

\item{offSet}{an integer. Default is 0L for no offset.}

\item{limit}{an integer. Optional.}
}
\description{
This adds a single sorted column to an ActivityInfo remote records object.
Usually one would use the \link[dplyr:slice]{dplyr::slice_head} or \link[dplyr:slice]{dplyr::slice_tail} verbs
instead.
}
