% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{admin.level.form.id}
\alias{admin.level.form.id}
\title{Returns the id of the form using the value of a 'classic' admin level form id}
\usage{
admin.level.form.id(adminLevelId)
}
\arguments{
\item{adminLevelId}{The admin level ID}
}
\description{
Returns the id of the form using the value of a 'classic' admin level form id
}
