% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{adminPermissions}
\alias{adminPermissions}
\title{adminPermissions}
\usage{
adminPermissions(
  manage_automations = FALSE,
  manage_users = FALSE,
  manage_roles = FALSE
)
}
\arguments{
\item{manage_automations}{Manage automations.}

\item{manage_users}{Manage database users.}

\item{manage_roles}{Assign roles to users.}
}
\description{
Helper method to create a list of database permissions for an administrative role.
}
\details{
Each argument may either be TRUE or FALSE.
}
