% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{as.data.frame.formField}
\alias{as.data.frame.formField}
\title{Flatten form field to a table}
\usage{
\method{as.data.frame}{formField}(element, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{element}{a \code{formField} element as found in a form schema.}

\item{...}{additional arguments passed on to \code{\link{data.frame}}.}

\item{stringsAsFactors}{should character vectors be converted to factors?}
}
\description{
Flatten form field to a table
}
