% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{as.data.frame.formSchema}
\alias{as.data.frame.formSchema}
\title{Flatten form schema as a table}
\usage{
\method{as.data.frame}{formSchema}(form)
}
\arguments{
\item{form.schema}{A list returned by \code{getFormSchema}.}
}
\description{
Flatten form schema as a table
}
