% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{asActivityDataFrame}
\alias{asActivityDataFrame}
\title{asActivityDataFrame}
\usage{
asActivityDataFrame(databaseSchema)
}
\description{
Creates a data.frame containing a list of activities
and their properties from the given database schema
}
