% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{asAttributeGroupDataFrame}
\alias{asAttributeGroupDataFrame}
\title{asAttributeGroupDataFrame}
\usage{
asAttributeGroupDataFrame(databaseSchema)
}
\description{
Creates a data.frame containing a list of attribute groups
and their properties from the given database schema
}
