% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{asIndicatorDataFrame}
\alias{asIndicatorDataFrame}
\title{asIndicatorDataFrame}
\usage{
asIndicatorDataFrame(databaseSchema)
}
\description{
Creates a data.frame containing a list of indicators
and their properties from the given database schema
}

