% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{attribute.field.name}
\alias{attribute.field.name}
\title{Returns the name of the field containing the value of a 'classic' attribute group}
\usage{
attribute.field.name(attributeGroupId)
}
\arguments{
\item{attributeGroupId}{The legacy ID}
}
\description{
Returns the name of the field containing the value of a 'classic' attribute group
}
