\name{authenticate}
\alias{authenticate}
\title{Constructs a httr::authentication object from saved credentials
from the user's home directory at ~/.activityinfo.credentials}
\usage{
authenticate()
}
\description{
Constructs a httr::authentication object from saved
credentials from the user's home directory at
~/.activityinfo.credentials
}

