% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cloneDatabase}
\alias{cloneDatabase}
\title{Creates copies of all existing activities in another
database}
\usage{
cloneDatabase(sourceDatabaseId, targetDatabaseId)
}
\description{
Creates copies of all existing activities in another
database
}

