% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copyPartners}
\alias{copyPartners}
\title{Copies the partners linked to 'sourceDatabase' to each database
in the vector 'targetDatabase'. Databases are identified by their numeric
identifier and you must, of course, have access to all databases.}
\usage{
copyPartners(sourceDatabase, targetDatabase)
}
\arguments{
\item{sourceDatabase}{the id of the source database}

\item{targetDatabase}{the id of the target database}
}
\description{
Copies the partners linked to 'sourceDatabase' to each database
in the vector 'targetDatabase'. Databases are identified by their numeric
identifier and you must, of course, have access to all databases.
}

