% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{copySchema}
\alias{copySchema}
\title{Create a new form schema copied from each form field represented in a remote 
records object}
\usage{
copySchema(x, databaseId, label, ...)
}
\arguments{
\item{x}{the remote records object fetched with getRecords().}

\item{databaseId}{the id of the database where the form should reside.}

\item{label}{the label of the form}
}
\description{
This function is useful for creating new forms based on existing forms and 
may be used to copy only the fields in a single form or any combination of 
fields in a form tree that can be fetched with getRecords(). Note some form 
elements such as section headers will not be represented.
}
