\name{createActivity}
\alias{createActivity}
\title{Creates a new activity}
\usage{
createActivity(databaseId, name, category = NULL, locationTypeId,
  reportingFrequency = "once")
}
\arguments{
  \item{databaseId}{the id of the database in which to
  create this activity}

  \item{name}{the activity's name}

  \item{category}{the new activity's category}

  \item{locationTypeId}{the id of the activity's location
  type}

  \item{reportingFrequency}{either "once" or "monthly"}
}
\description{
Creates a new activity
}

