% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{createDatabase}
\alias{createDatabase}
\title{createDatabase}
\usage{
createDatabase(label, databaseId = cuid())
}
\arguments{
\item{label}{The new database label}

\item{databaseId}{The new database identifier; a cuid will be created if left empty}
}
\description{
Creates a new database.
}
\examples{
\dontrun{
newDb <- createDatabase("Programme information system")
}
}
