% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{createIndicator}
\alias{createIndicator}
\title{Creates a new Indicator}
\usage{
createIndicator(activityId, name, category = NULL, listHeader = NULL,
  description = NULL, type = "quantity", units,
  aggregation = c("sum", "mean", "count"), sortOrder = NULL)
}
\arguments{
\item{activityId}{the numeric id of the activity to which this indicator
should be added}

\item{name}{the name of this indicator}

\item{category}{the activity's category as a string}

\item{listHeader}{a short label to be used when displaying the indicator as a
column}

\item{description}{an extended description of this indicator}

\item{units}{the units of measure of this indicator}

\item{aggregation}{the method to use when aggregating this indicator's values; 
can be "sum" (default), "mean" or "count"}

\item{sortOrder}{an ordinal value indicating the order of this indicator}
}
\value{
Returns the numeric id of the newly created indicator.
}
\description{
Creates a new Indicator
}
