% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuid.R
\name{cuid}
\alias{cuid}
\title{Generates a CUID}
\usage{
cuid(domain, id)
}
\description{
CUIDs are Collision-resistant IDs that are generated by clients and used to uniquely
identify forms, records, fields and other resources in ActivityInfo.
}
