% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{defaultColumnStyle}
\alias{defaultColumnStyle}
\title{Set or get the default column style as an R option}
\usage{
defaultColumnStyle(style)
}
\arguments{
\item{style}{the style object to set as the default style for column created
with a helper function such as idColumnStyle(), prettyColumnStyle(),
minimalColumnStyle() or allColumnStyle().}
}
\description{
For example to set a column style, the following are possible:

defaultColumnStyle(minimalColumnStyle())
defaultColumnStyle(allColumnStyle())
defaultColumnStyle(prettyColumnStyle(allReferenceFields = TRUE))

Use defaultColumnStyle() without an argument to get the current default style
object.
}
