% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{deleteAllActivitiesInDatabase}
\alias{deleteAllActivitiesInDatabase}
\title{Deletes all activities within a given database}
\usage{
deleteAllActivitiesInDatabase(databaseId)
}
\description{
Deletes all activities within a given database
}

