% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{deleteDatabase}
\alias{deleteDatabase}
\title{deleteDatabase}
\usage{
deleteDatabase(databaseId)
}
\arguments{
\item{databaseId}{database identifier}
}
\description{
Deletes a database.
}
\examples{
\dontrun{
deleteDatabase(databaseId = "c10011c3x5pnoldk0ua1qr")
}
}
