% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{deleteDatabaseUser}
\alias{deleteDatabaseUser}
\title{deleteDatabaseUser}
\usage{
deleteDatabaseUser(databaseId, userId)
}
\arguments{
\item{databaseId}{the id of the database}

\item{userId}{the (numeric) id of the user to remove from the database.}
}
\description{
Deletes a user from a database
}
