% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{deleteRecord}
\alias{deleteRecord}
\title{Delete a record}
\usage{
deleteRecord(formId, recordId)
}
\arguments{
\item{formId}{a form id}

\item{recordId}{a record id}
}
\description{
This call deletes a single record for the given \code{formId} and
\code{recordId}.
}
\examples{
\dontrun{

# Deletes a record
deleteRecord(formId = "cyx123", recordId = "c23g322j432")

# Recover (undelete) a record
recoverRecord(formId = "cyx123", recordId = "c23g322j432")
}
}
\seealso{
Other record functions: 
\code{\link{addRecord}()},
\code{\link{getAttachment}()},
\code{\link{getRecord}()},
\code{\link{getRecordHistory}()},
\code{\link{recoverRecord}()},
\code{\link{updateRecord}()}
}
\concept{record functions}
