% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{extractDatabaseLong}
\alias{extractDatabaseLong}
\title{Extract a full database in "long" format}
\usage{
extractDatabaseLong(database.id = NA)
}
\description{
Extract a full database in "long" format
}

