% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{form_database_record_table}
\alias{form_database_record_table}
\title{Record table description section}
\description{
Record table description section
}
\details{
\itemize{
  \item Columns for all Form Dimensions, including:
    \itemize{
      \item Sub-Form Name, Sub-Form Id, Sub-Form Record Id and Sub-Form Period (where applicable)
      \item Reference Field Dimensions (where applicable)
    }
  \item Rows for every measured value, with the measure field name under \dQuote{Measure} and the measured value under \dQuote{Value}
}
}
