% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{formSchema}
\alias{formSchema}
\title{Create a blank form schema offline}
\usage{
formSchema(
  databaseId,
  label,
  id = cuid(),
  elements = list(),
  folderId = databaseId
)
}
\arguments{
\item{databaseId}{The identifier of the database containing the form}

\item{label}{The label of the form}

\item{id}{The identifier of the form; if unused will generate a new cuid}

\item{elements}{The elements/form fields of the form}

\item{folderId}{The identifier of the folder containing the form}
}
\description{
Generates a new form schema offline which can be used to build a new form
}
