% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{generateId}
\alias{generateId}
\title{Generates a unique 31-bit integer id}
\usage{
generateId(n = 1)
}
\arguments{
\item{n}{the number of ids to generate}
}
\description{
Generates a unique 31-bit integer id
}

