% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{geopointFieldSchema}
\alias{geopointFieldSchema}
\title{Create a Geographic Point form field schema}
\usage{
geopointFieldSchema(
  label,
  description = NULL,
  requiredAccuracy = NULL,
  manualEntryAllowed = TRUE,
  code = NULL,
  id = cuid(),
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRule = "",
  validationRule = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{requiredAccuracy}{Minimum required accuracy in metres/meters.}

\item{manualEntryAllowed}{Whether coordinates may be added manually; default
is TRUE}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{required}{Whether the form field is required; default is FALSE}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRule}{Relevance rules for the form field given as a single character string; default is ""}

\item{validationRule}{Validation rules for the form field given as a single character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
A Geographic Point field allow users to enter a geo-location with a certain
accuracy.
}
\details{
The user interface as the required geo-location accuracy options:
* No minimum accuracy
* Medium accuracy (25 m); default
* High accuracy (10 m)"

This field cannot be a key field.
}
