% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{getActivityChangeLog}
\alias{getActivityChangeLog}
\title{Retrieves a data.frame containing all changes made to a given activity}
\usage{
getActivityChangeLog(activityId)
}
\description{
Retrieves a data.frame containing all changes made to a given activity
}
