% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAdminLevelEntities}
\alias{getAdminLevelEntities}
\title{Get all entities belonging to an administrative level by ID}
\usage{
getAdminLevelEntities(id)
}
\description{
Retrieves all entities belonging to an administrative level with the given
identifier.
}

