% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{getAdminLevels}
\alias{getAdminLevels}
\title{Get administrative levels within a country}
\usage{
getAdminLevels(country)
}
\arguments{
\item{country}{the country's numeric id or an ISO-3166 alpha-2 code (like
'US' or 'SS' or 'RW')}
}
\description{
Get administrative levels within a country
}

