% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingAccount}
\alias{getBillingAccount}
\title{getBillingAccount}
\usage{
getBillingAccount(asDataFrame = TRUE, billingId)
}
\arguments{
\item{asDataFrame}{Output as data.frame, Default: TRUE}

\item{billingId}{Billing ID}
}
\value{
Billing account information in list or data.frame output
}
\description{
Get billing account information
}
