% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingAccountDomains}
\alias{getBillingAccountDomains}
\title{getBillingAccountDomains}
\usage{
getBillingAccountDomains(billingId)
}
\arguments{
\item{billingId}{Billing ID}
}
\value{
Information on billing account email domain in list output
}
\description{
Billing account email domain info
}
