% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingDatabases}
\alias{getBillingDatabases}
\title{getBillingDatabases}
\usage{
getBillingDatabases(asDataFrame = TRUE, userId)
}
\arguments{
\item{asDataFrame}{Output as data.frame, Default: TRUE}

\item{userId}{User ID}
}
\value{
Information on databases under billing account in list or data.frame output
}
\description{
Data for all databases under billing account
}
