% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingDatabaseUsers}
\alias{getBillingDatabaseUsers}
\title{getBillingDatabaseUsers}
\usage{
getBillingDatabaseUsers(asDataFrame = TRUE, userId, databaseId)
}
\arguments{
\item{asDataFrame}{Data.frame output, Default: TRUE}

\item{userId}{User ID}

\item{databaseId}{Database ID}
}
\value{
User information from the specified database in list or data.frame output
}
\description{
Get data for users from a specific database. Can be more useful than \code{getDatabaseUsers()} as you also retrieve the database owner's info as well.
}
