% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingDomains}
\alias{getBillingDomains}
\title{getBillingDomains}
\usage{
getBillingDomains(userId)
}
\arguments{
\item{userId}{User ID}
}
\value{
Information on billing account email domain in list output
}
\description{
Billing account email domain info
}
