% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingInfo}
\alias{getBillingInfo}
\title{getBillingInfo}
\usage{
getBillingInfo(asDataFrame = TRUE, userId)
}
\arguments{
\item{asDataFrame}{Output as data.frame, Default: TRUE}

\item{userId}{User ID}
}
\value{
Billing account information in list or data.frame output
}
\description{
Get billing account information
}
