% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingUsers}
\alias{getBillingUsers}
\title{getBillingUsers}
\usage{
getBillingUsers(asDataFrame = TRUE, userId)
}
\arguments{
\item{asDataFrame}{Output as data.frame, Default: TRUE}

\item{userId}{User ID}
}
\value{
Billing account user(s) in list or data.frame output
}
\description{
Billing account users
}
