\name{getCube}
\alias{getCube}
\title{getCube}
\usage{
getCube(filters = list(), dimensions = c("indicator", "site"))
}
\arguments{
  \item{filters}{a named list of dimensions on which to
  filter}

  \item{dimensions}{a list of dimensions on which to
  roll-up}
}
\description{
Aggregates indicator values up to one or more dimensions.
}

