% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getDatabaseOwner}
\alias{getDatabaseOwner}
\title{getDatabaseOwner}
\usage{
getDatabaseOwner(asDataFrame = TRUE, databaseId)
}
\arguments{
\item{asDataFrame}{Data.frame output, Default: TRUE}

\item{databaseId}{Database ID}
}
\value{
Database owner in list output
}
\description{
Get database owner for a given database. This gives the owning organization rather than the user who created the database.
}
