% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseResources}
\alias{getDatabaseResources}
\title{getDatabaseResources}
\usage{
getDatabaseResources(databaseTree)
}
\arguments{
\item{databaseTree}{Database tree}
}
\description{
Creates a data.frame of database resources, types, parentIds and ids. 
This can be used to access a list of folders, forms, and sub-forms.
}
\examples{
\dontrun{
dbTree <- getDatabaseTree("c9mudk52")
dbResources <- getDatabaseResources(dbTree)
folders <- dbResources[dbResources$type == "FOLDER",]
forms <- dbResources[dbResources$type == "FORM",]
subForms <- dbResources[dbResources$type == "SUB_FORM",]
}

}
