% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseRoles}
\alias{getDatabaseRoles}
\title{getDatabaseRoles}
\usage{
getDatabaseRoles(database)
}
\arguments{
\item{database}{database tree using \link{getDatabaseTree} or the databaseId}
}
\description{
Get database roles in a data frame.
}
\examples{
\dontrun{
dbTree <- getDatabaseTree(databaseId = "ck3pqrp9a1z") # fetch the database tree
roles <- getDatabaseRoles(dbTree) # get the database roles
}
}
