% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabases}
\alias{getDatabases}
\title{getDatabases()}
\usage{
getDatabases(asDataFrame = TRUE)
}
\arguments{
\item{asDataFrame}{Retrieve database list as a data.frame, otherwise returns as list. Default: TRUE}
}
\description{
Retrieves a list of databases the authenticated user owns, or has been shared
with the authenticated user as a tibble. You can retrieve it as a list
if you set the argument \code{asDataFrame} to \code{FALSE}.
}
