% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{getDatabaseSchema}
\alias{getDatabaseSchema}
\title{getDatabaseSchema}
\usage{
getDatabaseSchema(databaseId)
}
\arguments{
\item{databaseId}{database identifier}
}
\description{
Retrieves the schema (partners, activities, indicators and attributes) from
for the given database.
}
