% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseTree}
\alias{getDatabaseTree}
\title{getDatabaseTree}
\usage{
getDatabaseTree(databaseId)
}
\arguments{
\item{databaseId}{database identifier}
}
\description{
Retrieves the database's tree of resources that are visible to the authenticated
user.
}
\examples{
\dontrun{
getDatabaseTree("ck2k93muu2")
}
}
