% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseUser}
\alias{getDatabaseUser}
\title{getDatabaseUser}
\usage{
getDatabaseUser(databaseId, userId)
}
\arguments{
\item{databaseId}{The database ID}

\item{userId}{The user ID}
}
\description{
Retrieves a user's role and permissions. Returns a NULL value if there is no user with the corresponding IDs.
}
