% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseUser2}
\alias{getDatabaseUser2}
\title{getDatabaseUser2}
\usage{
getDatabaseUser2(databaseId, userId)
}
\arguments{
\item{databaseId}{The database ID}

\item{userId}{The user ID}
}
\description{
Retrieves a user's role and permissions. This will throw an error if no user is found instead of returning a NULL value.
}
