% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseUsers}
\alias{getDatabaseUsers}
\title{getDatabaseUsers}
\usage{
getDatabaseUsers(databaseId)
}
\arguments{
\item{databaseId}{The database ID}
}
\description{
Retrieves the list of users with access to the database.
}
