% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{getDatabaseValueTable}
\alias{getDatabaseValueTable}
\title{Extract a full database in "long" format}
\usage{
getDatabaseValueTable(database.id = NA, include.comments = FALSE)
}
\description{
The resulting data.frame includes a row for each reported indicator
and month, with columns for the site's location, the administrative levels,
and for each of the single-valued attributes defined.
}

