% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{getFormData}
\alias{getFormData}
\title{getFormData}
\usage{
getFormData(activity, adminlevels, include.comments)
}
\arguments{
\item{activity}{activity}

\item{adminlevels}{adminlevels}

\item{include.comments}{include.comments}
}
\description{
getFormData
}
