% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{getFormRecordTable}
\alias{getFormRecordTable}
\title{Extract all Form Records, including all Sub-Form Records, in "long" format.}
\usage{
getFormRecordTable(form.id = NA, col.names = NULL)
}
\arguments{
\item{form.id}{the full alphanumeric id of the form}

\item{col.names}{a named character vector containing alternate names for the resulting table}
}
\description{
The Form Record Table is returned as a data.frame, with:
-> Columns for all Form Dimensions, including:
   - Sub-Form Name, Sub-Form Id, Sub-Form Record Id and Sub-Form Period (where applicable)
   - Reference Field Dimensions (where applicable)
-> Rows for every measured value, with the measure field name under "Measure" and the measured value under "Value"
}
