% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{getFormTree}
\alias{getFormTree}
\title{Queries the Form Tree of a Form}
\usage{
getFormTree(formId)
}
\arguments{
\item{formId}{the form identifier}
}
\value{
A list with class \sQuote{formTree}.
}
\description{
Queries the Form Tree of a Form
}
