% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLocationTypes}
\alias{getLocationTypes}
\title{Get location types within a country}
\usage{
getLocationTypes(country)
}
\arguments{
\item{country}{the country's numeric id or an ISO-3166 alpha-2 code (like 'US' or 'SS' or 'RW')}
}
\description{
Get location types within a country
}

