% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{getNextExpansions}
\alias{getNextExpansions}
\title{Get next possible expansions from an ActivityInfo variable expression or form id}
\usage{
getNextExpansions(x, path = NULL)
}
\arguments{
\item{x}{An ActivityInfo get records object, form id, or form tree.}

\item{path}{A variable expression (path) or form ID.}
}
\value{
A character vector of possible next expansions (field labels, IDs, or codes).
}
\description{
A simple helper function to get the variables/columns available from a form
or reference field.
}
\examples{
\dontrun{

x <- getRecords(formId)

# Get a list of fields available from the "Office administrator" reference 
# form, e.g. from the Staff form
available_fields <- getNextExpansions(x, "[Office administrator]")

}

}
