% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPartnersDataFrame}
\alias{getPartnersDataFrame}
\title{Fetch the partners registered in a database as a data.frame}
\usage{
getPartnersDataFrame(databaseId)
}
\arguments{
\item{databaseId}{database identifier}
}
\value{
A data frame with columns \code{id} and \code{name}. This data frame
is empty if no partners are registered in the database.
}
\description{
Fetch the partners registered in a database as a data.frame
}

