% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getRecordHistory}
\alias{getRecordHistory}
\title{Gets the list of changes to a record}
\usage{
getRecordHistory(formId, recordId, asDataFrame = TRUE)
}
\arguments{
\item{formId}{a form id}

\item{recordId}{a record id}

\item{asDataFrame}{Retrieve user list as a data.frame, otherwise returns as list. Default: TRUE}
}
\description{
This calls retrieves a list of all changes to the record, and the users
who initially added and then have modified that record.
}
\seealso{
Other record functions: 
\code{\link{addRecord}()},
\code{\link{deleteRecord}()},
\code{\link{getAttachment}()},
\code{\link{getRecord}()},
\code{\link{recoverRecord}()},
\code{\link{updateRecord}()}
}
\concept{record functions}
