% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getResource}
\alias{getResource}
\title{Get a resource by path}
\usage{
getResource(path, queryParams = list(...), ...)
}
\description{
Retrieves a single resource from the given path.
Remaining arguments are treated as query parameters
and must be named
}

