% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{getSiteData}
\alias{getSiteData}
\title{Fetch the sites for an activity as a data.frame}
\usage{
getSiteData(activity, adminlevels, include.comments)
}
\description{
Fetch the sites for an activity as a data.frame
}
