% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{grant}
\alias{grant}
\title{Create a grant to define resource-level permissions}
\usage{
grant(resourceId, permissions = resourcePermissions(), optional = FALSE)
}
\arguments{
\item{resourceId}{the id of the resource, for example a database, folder or
(sub-)form}

\item{permissions}{a permission list; see \link{resourcePermissions}}

\item{optional}{whether the grant is optional, by default it is not optional
(=FALSE)}
}
\description{
Returns a grant that can be added to a \link{role}.
}
\details{
Grants define access to resources such as databases, folders, or forms. The
permissions include operations such as view, read or edit and are defined per
resource. See \link{resourcePermissions}.

Adding grants to a role enables the administrator to define
permissions that vary per grant and, if desired, override grants inherited
from parent resources, such as a folder.

A grant can be set as optional, which means that you can choose whether to
enable the grant for each user that you invite to your database.

See \link{role} for the creation of roles.
}
\examples{
\dontrun{

optionalGrant <- 
  grant(resourceId = "ck5dxt1552",
      permissions = resourcePermissions(
        view = TRUE,
        add_record = TRUE,
        edit_record = TRUE
        ),
      optional = TRUE
)
}
}
