% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{idColumnStyle}
\alias{idColumnStyle}
\title{A form table style including all columns with ids as column names}
\usage{
idColumnStyle(maxDepth = 2)
}
\arguments{
\item{maxDepth}{the maximum depth of recursion in referenced forms. This is set to 3 by default.}
}
\description{
This function is primarily used in \link{getRecords}.
}
