% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{indicator.field.name}
\alias{indicator.field.name}
\title{Returns the name of the field containing the value of a 'classic' indicator}
\usage{
indicator.field.name(indicatorId)
}
\description{
Returns the name of the field containing the value of a 'classic' indicator
}
