% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{is.activity}
\alias{is.activity}
\title{Returns TRUE if the given object is an activity
object included from getDatabaseSchema}
\usage{
is.activity(activity)
}
\description{
Returns TRUE if the given object is an activity
object included from getDatabaseSchema
}

