% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{migrateFieldData}
\alias{migrateFieldData}
\title{Migrate and convert the data of one form field into another}
\usage{
migrateFieldData(
  .data,
  from,
  to,
  fn = function(x) x,
  idColumn = as.name("_id")
)
}
\arguments{
\item{.data}{remote records object of the form online}

\item{from}{the source form field from which to get the data}

\item{to}{the destination form field which will receive the converted data}

\item{fn}{the user-supplied R conversion function; default is to do nothing}

\item{idColumn}{the id column. The default is \verb{_id}}
}
\value{
The remote records object.
}
\description{
With this function, the data from one form field (column) can be moved to
another form field and converted with a user-supplied R function.
}
