% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{migrateFormField}
\alias{migrateFormField}
\alias{migrateFormField.formSchema}
\title{Migrate and convert with a user function the content of one form field to 
another}
\usage{
migrateFormField(...)

\method{migrateFormField}{formSchema}(formSchema, fn)
}
\arguments{
\item{...}{ignored}

\item{formSchema}{The offline schema of the form}

\item{formId}{The identifier of the form online}

\item{schema}{The form field schema to be added to the form}

\item{upload}{Default is FALSE. If TRUE the modified form schema will be uploaded.}
}
\value{
The form field schema after the addition This will be the form field schema from the server if changes are uploaded.
}
\description{
With this function, the data from one form field (column) can be moved to 
another form field and converted with a user-supplied function.
}
