% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{minimalColumnStyle}
\alias{minimalColumnStyle}
\title{A form table style with columns limited to those in the web interface}
\usage{
minimalColumnStyle()
}
\description{
This function is primarily used in getRecords().
}
