% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{monthly.reports.form.id}
\alias{monthly.reports.form.id}
\title{Returns the id of the form containing the monthly reports
associated with a 'classic' activity}
\usage{
monthly.reports.form.id(activityId)
}
\arguments{
\item{activityId}{The activity ID}
}
\description{
Returns the id of the form containing the monthly reports
associated with a 'classic' activity
}
