% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{noteFieldSchema}
\alias{noteFieldSchema}
\title{Create a note form field schema}
\usage{
noteFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  hideFromEntry = FALSE,
  relevanceRule = ""
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The text of the note}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{relevanceRule}{Relevance rules for the form field given as a single character string; default is ""}
}
\description{
A note displays its label and description to users during data entry. It
does not capture a value and is not included in the records table.
}
\seealso{
Other field schemas: 
\code{\link{attachmentFieldSchema}()},
\code{\link{barcodeFieldSchema}()},
\code{\link{calculatedFieldSchema}()},
\code{\link{dateFieldSchema}()},
\code{\link{formFieldSchema}()},
\code{\link{geopointFieldSchema}()},
\code{\link{monthFieldSchema}()},
\code{\link{multilineFieldSchema}()},
\code{\link{multipleReferenceFieldSchema}()},
\code{\link{multipleSelectFieldSchema}()},
\code{\link{quantityFieldSchema}()},
\code{\link{referenceFieldSchema}()},
\code{\link{sectionFieldSchema}()},
\code{\link{serialNumberFieldSchema}()},
\code{\link{singleSelectFieldSchema}()},
\code{\link{subformFieldSchema}()},
\code{\link{userFieldSchema}()},
\code{\link{weekFieldSchema}()}
}
\concept{field schemas}
