% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{parameter}
\alias{parameter}
\title{Create a role parameter to add to a user role definition}
\usage{
parameter(id, label, range)
}
\arguments{
\item{id}{the id of the parameter, for example "partner", which can
be used in a formula as "@user.partner". The id starts with a letter and may
contain letters, numbers and underscores _ under 32 characters.}

\item{label}{the label of the partner, for example, "Reporting partner"}

\item{range}{the id of a reference table, for example the list of partners,
or a formula}
}
\description{
Returns a role parameter.
}
\details{
Parameters are used to set up conditions that can be defined per user when
the role is given to a user or a user is created. A common use-case is to
restrict the user to only edit records related to the reporting partner for
which they work or only the region for which they are responsible. A
parameter enables the administrator to set their organization or region when
giving them a role.

See \link{role} for the creation of roles.
}
\examples{
\dontrun{

parameter(id = "partner", label = "Reporting partner", range = "ck5dxt1712")
}
}
