% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{permissions}
\alias{permissions}
\title{permissions}
\usage{
permissions(
  view = TRUE,
  add_record = FALSE,
  edit_record = FALSE,
  delete_record = FALSE,
  export_records = FALSE,
  lock_records = FALSE,
  add_resource = FALSE,
  edit_resource = FALSE,
  delete_resource = FALSE,
  manage_collection_links = FALSE,
  manage_users = FALSE,
  manage_roles = FALSE,
  manage_reference_data = FALSE,
  audit = FALSE,
  share_reports = FALSE,
  publish_reports = FALSE
)
}
\arguments{
\item{view}{View the resource, whether a form, folder, or database.}

\item{add_record}{Add a record within a form contained by this folder or form.}

\item{edit_record}{Edit a record's values within a form contained by this folder or form.}

\item{delete_record}{Delete a record within this form.}

\item{export_records}{Export Records from a form, folder or database.}

\item{lock_records}{Add, modify, or remove locks on records.}

\item{add_resource}{Create a new Resource (Form or Folder).}

\item{edit_resource}{Edit a Resource's schema, structure, attributes or data.}

\item{delete_resource}{Delete a Resource (Form or Folder).}

\item{manage_collection_links}{Manage (open/close) collection links for the given form.}

\item{manage_users}{Grant permissions to a user to this database, folder, or form.}

\item{manage_roles}{Add, modify and delete roles.}

\item{manage_reference_data}{Manage reference data.}

\item{audit}{Access the Audit logs for a database (or a subset).}

\item{share_reports}{Allow the user to share reports with other roles in the database.}

\item{publish_reports}{Allows the user to publish reports.}
}
\description{
Helper method to create a list of permissions for a role or grant.
}
\details{
Each argument may either be TRUE or FALSE.

The view, add_record, edit_record, and delete_record permissions may instead
be a formula that conditions the permission on the values of the record.
}
