% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{prettyColumns}
\alias{prettyColumns}
\title{Create the column selection object in the pretty style used by the web
interface including record metadata for queryTable()}
\usage{
prettyColumns(x, select, ...)
}
\arguments{
\item{x}{the form id, form schema, form tree, or remote records object.}

\item{select}{a character vector of column names to select.}

\item{...}{parameters passed on to prettyColumnStyle()}
}
\description{
This function is primarily used for queryTable().
}
