% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{print.activityInfoFormFieldSchema}
\alias{print.activityInfoFormFieldSchema}
\title{Pretty print a form field schema}
\usage{
\method{print}{activityInfoFormFieldSchema}(x, ...)
}
\arguments{
\item{x}{an object of class \code{activityInfoFormFieldSchema}.}

\item{...}{ignored}
}
\description{
Pretty print a form field schema
}
