% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{print.formSchema}
\alias{print.formSchema}
\title{Pretty print a form schema}
\usage{
\method{print}{formSchema}(x, ...)
}
\arguments{
\item{x}{an object of class \code{formSchema}.}

\item{...}{ignored}
}
\description{
Pretty print a form schema
}
