% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{queryReportResults}
\alias{queryReportResults}
\title{Query a pivot table report}
\usage{
queryReportResults(reportId)
}
\arguments{
\item{reportId}{The report idenitifier}
}
\value{
A data frame with the pivot table data.
}
\description{
Query a pivot table report
}
