% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{recordExists}
\alias{recordExists}
\title{Checks whether a record exists}
\usage{
recordExists(formId, recordId)
}
\arguments{
\item{the}{id of the record to check}
}
\description{
Checks whether a record exists
}
