% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{reference}
\alias{reference}
\title{Creates a 'reference' field value}
\usage{
reference(formId, recordId)
}
\arguments{
\item{formId}{the id of the form}

\item{recordId}{the id of the record}
}
\description{
Creates a 'reference' field value
}
