% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{relocateForm}
\alias{relocateForm}
\title{Relocating a form moves the form, together with all of its subforms,
records, and record history, to a new database.}
\usage{
relocateForm(formId, newDatabaseId)
}
\arguments{
\item{formId}{the id of the form to move}

\item{databaseId}{the id of the database to which the form should be moved.}
}
\description{
In order to relocate a form, the user must have "delete" permission in the
source database, and permission to add new forms in the target database.
}
