% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{role}
\alias{role}
\title{Create a role}
\usage{
role(
  id,
  label,
  parameters = list(),
  grants,
  permissions = managementPermissions()
)
}
\arguments{
\item{id}{the id of the role}

\item{label}{the label or name of the role, e.g. "Viewer" or "Administrator"}

\item{parameters}{a list of \link{parameter} items defining role parameters}

\item{grants}{a list of \link{grant} items for each resources and their
respective permissions}

\item{filters}{a list of \link{roleFilter} items}

\item{managementPermissions}{\link{adminPermissions} under this role}
}
\description{
Returns a role that can be added to a database and assigned to users.
}
\details{
A role contains one or more \link{grant} items that define access to
resources (database, folder, forms).

Some administrative permissions are defined at the level of the role rather
than within grants. See \link{adminPermissions}.
}
\examples{
\dontrun{

# Create a Reporting Partner role that may view, and edit their own records 
in the form with id "cq9xyz1552" and optional access to view and add records
to a form with id "cz55555555". They are also allowed to discover these 
forms.

grantBasedRole <- 
  role(id = "rp",
      label = "Reporting Partner",
      parameters = list(
        parameter(id = "partner", label = "Partner", range = "ck5dxt1712")),
      grants = list(
        grant(resourceId = "cq9xyz1552",
          permissions = permissions(
            view = "ck5dxt1712 == @user.partner",
            edit_record = "ck5dxt1712 == @user.partner",
            discover = TRUE,
            export_records = TRUE)),
        grant(resourceId = "cz55555555",
          permissions = permissions(
            view = TRUE,
            discover = TRUE,
            add_record = TRUE),
          optional = TRUE))
      )
      
}
}
