% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{roleFilter}
\alias{roleFilter}
\title{Create a pre-defined legacy role filter}
\usage{
roleFilter(id, label, filter)
}
\arguments{
\item{id}{the id of the pre-defined filter}

\item{label}{A human-readable label}

\item{filter}{A formula that can be used to filter a record-level permission.}
}
\description{
This should no longer be used. Please update your code to use grant-based
roles. This function is deprecated.
}
\details{
Pre-defined filters. Role filters allow other users to choose filters for
permissions without having to write formulas themselves. This is a feature of
legacy roles.
}
\examples{
\dontrun{

legacyRoleFilter <- 
  roleFilter(
          id = "partner", 
          label = "Partner is user's partner", 
          filter = "ck5dxt1712 == @user.partner"
          )
}
}
