% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{serialNumberFieldSchema}
\alias{serialNumberFieldSchema}
\title{Create a serial number form field schema object}
\usage{
serialNumberFieldSchema(
  label,
  description = NULL,
  digits = 5L,
  prefixFormula = NULL,
  code = NULL,
  id = cuid(),
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{digits}{The number of digits in the serial number}

\item{prefixFormula}{A formula as a character string defining the prefix for
the serial number}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
Only one serial number field is possible in a form. The Prefix Formula is
available for Serial Number fields and can be used to customise how the
Serial Number will appear. In Subforms, the Prefix Formula can also derive
from the Parent Form.
}
\seealso{
Other field schemas: 
\code{\link{attachmentFieldSchema}()},
\code{\link{barcodeFieldSchema}()},
\code{\link{calculatedFieldSchema}()},
\code{\link{dateFieldSchema}()},
\code{\link{formFieldSchema}()},
\code{\link{geopointFieldSchema}()},
\code{\link{monthFieldSchema}()},
\code{\link{multilineFieldSchema}()},
\code{\link{multipleReferenceFieldSchema}()},
\code{\link{multipleSelectFieldSchema}()},
\code{\link{noteFieldSchema}()},
\code{\link{quantityFieldSchema}()},
\code{\link{referenceFieldSchema}()},
\code{\link{sectionFieldSchema}()},
\code{\link{singleSelectFieldSchema}()},
\code{\link{subformFieldSchema}()},
\code{\link{userFieldSchema}()},
\code{\link{weekFieldSchema}()}
}
\concept{field schemas}
