% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{site.form.id}
\alias{site.form.id}
\title{Returns the id of the form containing the sites
associated with a 'classic' activity}
\usage{
site.form.id(activityId)
}
\arguments{
\item{activityId}{The activity ID}
}
\description{
Returns the id of the form containing the sites
associated with a 'classic' activity
}
