% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{stageImport}
\alias{stageImport}
\title{Stages data to import to ActivityInfo}
\usage{
stageImport(text)
}
\arguments{
\item{text}{The text of the file to import.}
}
\description{
Stages data to import to ActivityInfo
}
