% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{subformFieldSchema}
\alias{subformFieldSchema}
\title{Create a Subform field schema}
\usage{
subformFieldSchema(
  label,
  description = NULL,
  subformId,
  code = NULL,
  id = cuid(),
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRule = "",
  validationRule = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{subformId}{The id of the sub-form}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRule}{Relevance rules for the form field given as a single character string; default is ""}

\item{validationRule}{Validation rules for the form field given as a single character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
A subform field can be used to define a field that contains a subform.
}
\details{
A subform field cannot be a key field.
}
\seealso{
Other field schemas: 
\code{\link{attachmentFieldSchema}()},
\code{\link{barcodeFieldSchema}()},
\code{\link{calculatedFieldSchema}()},
\code{\link{dateFieldSchema}()},
\code{\link{formFieldSchema}()},
\code{\link{geopointFieldSchema}()},
\code{\link{monthFieldSchema}()},
\code{\link{multilineFieldSchema}()},
\code{\link{multipleReferenceFieldSchema}()},
\code{\link{multipleSelectFieldSchema}()},
\code{\link{quantityFieldSchema}()},
\code{\link{referenceFieldSchema}()},
\code{\link{reverseReferenceFieldSchema}()},
\code{\link{sectionFieldSchema}()},
\code{\link{serialNumberFieldSchema}()},
\code{\link{singleSelectFieldSchema}()},
\code{\link{userFieldSchema}()},
\code{\link{weekFieldSchema}()}
}
\concept{field schemas}
