% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending.R
\name{submitPending}
\alias{submitPending}
\title{Imports a pending change set exported from an ActivityInfo offline database.}
\usage{
submitPending(file.name)
}
\description{
If a device is no longer able to connect to the internet, you can export
the changes queued for synchronization and manually transfer the pending.json
file to the database owner, who can use this send the changes to the server
under their own account.
}
