% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{tbl_format_header.tbl_activityInfoRemoteRecords}
\alias{tbl_format_header.tbl_activityInfoRemoteRecords}
\title{#' @export
format.tbl_activityInfoRemoteRecords <- function(x) {
  format(collect(adjustWindow(x, offSet = 0L, limit = 2L)))
}}
\usage{
\method{tbl_format_header}{tbl_activityInfoRemoteRecords}(x, setup, ...)
}
\description{
#' @export
format.tbl_activityInfoRemoteRecords <- function(x) {
  format(collect(adjustWindow(x, offSet = 0L, limit = 2L)))
}
}
