% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{tblColumns}
\alias{tblColumns}
\title{Get the columns of a remote records object}
\usage{
tblColumns(x)
}
\arguments{
\item{x}{the remote records object fetched with getRecords().}
}
\description{
This provides the columns of the remote records object created with getRecords()
}
