% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{tblFilter}
\alias{tblFilter}
\title{Get the filter object of a remote records object}
\usage{
tblFilter(x)
}
\arguments{
\item{x}{the remote records object fetched with \link{getRecords}.}
}
\description{
This provides the columns used in \link{queryTable} of the remote records object created with \link{getRecords}
}
