% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{tblSort}
\alias{tblSort}
\title{Get the sort ActivityInfo API object of a remote records object}
\usage{
tblSort(x)
}
\arguments{
\item{x}{the remote records object fetched with \link{getRecords}.}
}
\description{
This provides the sort object used in \link{queryTable} of the remote records object created with \link{getRecords}
}
