% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{tblWindow}
\alias{tblWindow}
\title{Get the window object of a remote records object}
\usage{
tblWindow(x, limit)
}
\arguments{
\item{x}{the remote records object fetched with getRecords().}

\item{limit}{an additional limit that can be used to specify the maximum number of records to query.}
}
\description{
This provides the window used in queryTable()of the remote records object created with getRecords()
}
