% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{toActivityInfoJson}
\alias{toActivityInfoJson}
\title{Wraps jsonlite::toJSON to match the conventions expected by the
ActivityInfo server}
\usage{
toActivityInfoJson(x)
}
\arguments{
\item{x}{an R object to convert to JSON}
}
\description{
Wraps jsonlite::toJSON to match the conventions expected by the
ActivityInfo server
}
