% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{toSelectOptions}
\alias{toSelectOptions}
\alias{toSelectOptions.character}
\alias{toSelectOptions.default}
\alias{toSelectOptions.list}
\alias{toSelectOptions.factor}
\title{Converts a factor, list or vector to select options}
\usage{
toSelectOptions(options)

\method{toSelectOptions}{character}(options)

\method{toSelectOptions}{default}(options)

\method{toSelectOptions}{list}(options)

\method{toSelectOptions}{factor}(options)
}
\arguments{
\item{options}{The character vector or list to convert to options}
}
\description{
Creates a list of options in the ActivityInfo format for Single Select or
Multiple Select form fields.
}
