% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{toSelectValues}
\alias{toSelectValues}
\alias{toSelectValues.character}
\alias{toSelectValues.default}
\alias{toSelectValues.list}
\alias{toSelectValues.factor}
\title{Converts a factor, list or vector to select values}
\usage{
toSelectValues(values)

\method{toSelectValues}{character}(values)

\method{toSelectValues}{default}(values)

\method{toSelectValues}{list}(values)

\method{toSelectValues}{factor}(values)
}
\arguments{
\item{values}{The values to convert}
}
\description{
Creates a list of values in the ActivityInfo format for Single Select or
Multiple Select form fields.
}
