% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{updateFormSchema}
\alias{updateFormSchema}
\title{Updates a form schema}
\usage{
updateFormSchema(schema)
}
\arguments{
\item{schema}{a form schema}
}
\value{
Returns the updated form schema from the ActivityInfo server
}
\description{
Updates a form schema
}
