% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{updateRecord}
\alias{updateRecord}
\title{Updates a single record}
\usage{
updateRecord(formId, recordId, fieldValues)
}
\arguments{
\item{fieldValues}{a named list of fields to change}
}
\description{
Updates a single record
}
