% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{updateRole}
\alias{updateRole}
\title{Updates a role's definition in the database}
\usage{
updateRole(databaseId, role)
}
\arguments{
\item{databaseId}{the id of the database}

\item{role}{the role definition,}
}
\description{
Updates a role's definition in the database
}
\examples{
\dontrun{
updateRole("cxy123", list(

}

}
