% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{updateUserRole}
\alias{updateUserRole}
\title{updateUserRole}
\usage{
updateUserRole(databaseId, userId, assignment)
}
\arguments{
\item{databaseId}{the id of the database}

\item{userId}{the (numeric) id of the user to update}

\item{assignment}{the role assignment, \code{\link[activityinfo]{roleAssignment}}}
}
\description{
Updates a user's role in the database
}
\examples{
\dontrun{

databaseId <- "caxadcasdf"
updateUserRole(databaseId,
  userId = 165,
  roleAssignment(
    roleId = "admin",
    roleResources = databaseId
  )
)
}

}
