% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.r
\name{validateSerialNumberFieldSchema}
\alias{validateSerialNumberFieldSchema}
\title{Validate a serial number form field schema}
\usage{
validateSerialNumberFieldSchema(schema)
}
\arguments{
\item{schema}{The schema of the serial number form field}
}
\description{
Validate a serial number form field schema
}
