% Generated by roxygen2 (4.0.2): do not edit by hand
\name{activityInfoLogin}
\alias{activityInfoLogin}
\title{Configures the current session to use a user's email address and
password for authentication to ActivityInfo.org.}
\usage{
activityInfoLogin(userEmail, password)
}
\arguments{
\item{userEmail}{The email address used to log in with to ActivityInfo.org}

\item{password}{The user's ActivityInfo password}
}
\description{
Configures the current session to use a user's email address and
password for authentication to ActivityInfo.org.
}

