% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{activityInfoRootUrl}
\alias{activityInfoRootUrl}
\title{Get or set the root url for the ActivityInfo}
\usage{
activityInfoRootUrl(newUrl)
}
\arguments{
\item{newUrl}{The new URL to set as the ActivityInfo root URL}
}
\description{
This call gets or sets the root url used for a session, valid only
during this session.
}
\details{
If you are using the [ActivityInfo Self-Managed Server](https://www.activityinfo.org/support/docs/self-managed/index.html),
you can use this function to point the R package to your own server.
}
\examples{
# Connect to a self-managed ActivityInfo server
activityInfoRootUrl("https://activityinfo.example.org")

}
