% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{activityInfoRootUrl}
\alias{activityInfoRootUrl}
\title{Get or set the root url for the ActivityInfo}
\usage{
activityInfoRootUrl(new.url)
}
\description{
This call gets or sets the root url used for this session, \strong{valid only
during the session}.
}
