% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{activityInfoRootUrl}
\alias{activityInfoRootUrl}
\title{Get or set the root url for the ActivityInfo server}
\usage{
activityInfoRootUrl(newUrl)
}
\arguments{
\item{newUrl}{The new URL to set as the ActivityInfo root URL}
}
\description{
Get or set the root url for the ActivityInfo server
}
\details{
If you are using the \href{https://www.activityinfo.org/support/docs/self-managed/index.html}{ActivityInfo Self-Managed Server},
you can use this function to point the R package to your own server.
}
\examples{
\dontrun{
# Connect to a self-managed ActivityInfo server
activityInfoRootUrl("https://activityinfo.example.org")
}
}
