% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{addDatabaseUser}
\alias{addDatabaseUser}
\title{inviteUser}
\usage{
addDatabaseUser(
  databaseId,
  email,
  name,
  locale = NA_character_,
  roleId,
  roleParameters = list(),
  roleResources = list(databaseId)
)
}
\arguments{
\item{databaseId}{the id of the database to which they should be added}

\item{email}{the user's email}

\item{name}{the user's name (only used if they do not already have an ActivityInfo account)}

\item{locale}{the locale ("en', "fr", "ar", etc) to use inviting the user (only used if they do not already have an ActivityInfo account)}

\item{roleId}{the id of the role to assign to the user}

\item{roleParameters}{a named list containing the role parameter values}

\item{roleResources}{a list of folders in which this role should be assigned (or the databaseId if they should have this role in the whole database)}
}
\description{
Invites a user to a database.
}
