% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{addForm}
\alias{addForm}
\title{Adds a new form to a database}
\usage{
addForm(databaseId, schema, folderId = databaseId)
}
\arguments{
\item{databaseId}{the id of the database}

\item{schema}{the schema of the form to add}

\item{folderId}{the id of the folder to which this form should be added}
}
\description{
Adds a new form to a database
}
