% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.r
\name{addFormField}
\alias{addFormField}
\alias{addFormField.character}
\alias{addFormField.formSchema}
\alias{addFormField.default}
\title{Add a new form field}
\usage{
addFormField(...)

\method{addFormField}{character}(formId, schema, upload = FALSE, ...)

\method{addFormField}{formSchema}(formSchema, schema, upload = FALSE, ...)

\method{addFormField}{default}(formId, schema, upload = FALSE, ...)
}
\arguments{
\item{...}{ignored}

\item{formId}{The identifier of the form online}

\item{schema}{The form field schema to be added to the form}

\item{upload}{Default is FALSE. If TRUE the modified form schema will be uploaded.}

\item{formSchema}{The offline schema of the form}
}
\description{
Adds a new form field to an offline form schema or else downloads the form 
schema and adds the new form field. Note that the either the upload argument 
must be TRUE for the field to be added automatically online or the user will 
also need to use updateFormSchema() to upload the changes after they are 
completed.
}
