% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{addSort}
\alias{addSort}
\title{Add an ActivityInfo API style sort object}
\usage{
addSort(x, sort)
}
\arguments{
\item{x}{the remote records object fetched with getRecords().}

\item{sort}{the sort object in the following format: list(list(dir = "ASC", field = "[Name]")) or list(list(dir = "DESC", field = "[Name]"))}
}
\description{
This adds a single sorted column to an ActivityInfo remote records object. 
Usually one would use the dplyr::arrange() verb instead.
}
