% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{as.data.frame.formField}
\alias{as.data.frame.formField}
\title{Flatten form field to a table}
\usage{
\method{as.data.frame}{formField}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an object of class \emph{formField} element as found in a form schema.}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical, if \code{TRUE} then converting column names is optional.}

\item{...}{additional arguments passed on to \code{\link{as.data.frame}}.}
}
\description{
Flatten form field to a table
}
\details{
Note that if \code{stringsAsFactors} is not explicitly set to \code{TRUE}, then this method uses
\code{FALSE} as a default, not \code{default.stringsAsFactors()}.
}
