% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{as.data.frame.formSchema}
\alias{as.data.frame.formSchema}
\title{Flatten form schema to a table}
\usage{
\method{as.data.frame}{formSchema}(form, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{additional arguments passed on to \code{\link{as.data.frame}}.}

\item{stringsAsFactors}{should character vectors be converted to factors?}

\item{form.schema}{A list returned by \code{\link{getFormSchema}}.}
}
\description{
Flatten form schema to a table
}
