% Generated by roxygen2 (4.0.2): do not edit by hand
\name{asActivityDataFrame}
\alias{asActivityDataFrame}
\title{asActivityDataFrame}
\usage{
asActivityDataFrame(databaseSchema)
}
\description{
Creates a data.frame containing a list of activities
and their properties from the given database schema
}

