% Generated by roxygen2 (4.0.2): do not edit by hand
\name{asAttributeGroupDataFrame}
\alias{asAttributeGroupDataFrame}
\title{asAttributeGroupDataFrame}
\usage{
asAttributeGroupDataFrame(databaseSchema)
}
\description{
Creates a data.frame containing a list of attribute groups
and their properties from the given database schema
}

