% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.r
\name{barcodeFieldSchema}
\alias{barcodeFieldSchema}
\title{Create a barcode form field schema}
\usage{
barcodeFieldSchema(...)
}
\arguments{
\item{...}{See arguments defined in \link[activityinfo]{formFieldSchema}}
}
\description{
Create a barcode form field schema
}
