% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{calculatedFieldSchema}
\alias{calculatedFieldSchema}
\title{Create a calculated field schema}
\usage{
calculatedFieldSchema(
  label,
  description = NULL,
  formula,
  code = NULL,
  id = cuid(),
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRule = "",
  validationRule = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{formula}{A character string with the calculation formula}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRule}{Relevance rules for the form field given as a single character string; default is ""}

\item{validationRule}{Validation rules for the form field given as a single character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
A calculated field can be used to define a formula that presents calculation 
results in the form.
}
\details{
A calculated field cannot be a key field.
}
