% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{cloneDatabase}
\alias{cloneDatabase}
\title{Creates copies of all existing activities in another
database}
\usage{
cloneDatabase(sourceDatabaseId, targetDatabaseId)
}
\description{
Creates copies of all existing activities in another
database
}
