% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{createActivity}
\alias{createActivity}
\title{Creates a new activity}
\usage{
createActivity(databaseId, name, category = NULL, locationTypeId,
  reportingFrequency = c("once", "monthly"), ...)
}
\arguments{
\item{databaseId}{the numeric id of the database in which to create this
activity}

\item{name}{the activity's name}

\item{category}{the new activity's category as a string}

\item{locationTypeId}{the numeric id of the activity's location type}

\item{reportingFrequency}{a string that defines the reporting frequency; can 
be "once" (default) or "monthly"}

\item{...}{additional key=value arguments which are passed to the
createEntity function. For example: \code{classicView = TRUE} to create a
classic activity form.}
}
\value{
Returns the numeric id of the newly created activity.
}
\description{
Creates a new activity
}
