\name{createDatabase}
\alias{createDatabase}
\title{Creates a new database}
\usage{
createDatabase(name, description = NULL, countryId)
}
\arguments{
  \item{name}{a short-ish name for the database}

  \item{description}{a longer description of the database}

  \item{country}{the location and geographic reference data
  to use}
}
\description{
Creates a new database
}

