% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{createDatabase}
\alias{createDatabase}
\title{Creates a new database}
\usage{
createDatabase(name, description = NULL, countryId)
}
\arguments{
\item{name}{a short-ish name for the database}

\item{description}{a longer description of the database}

\item{country}{the location and geographic reference data to use: either the
numeric identifier used by ActivityInfo or an ISO-3166 two-letter code}
}
\value{
Returns the numeric id of the newly created database.
}
\description{
Creates a new database
}
\examples{
\dontrun{
id <- createDatabase(name="My Database in Bangladesh", countryId = "BD")
}
}
\references{
See \url{http://en.wikipedia.org/wiki/ISO_3166-2} for two-letter
  country codes.
}

