% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{createFormSchemaFromData}
\alias{createFormSchemaFromData}
\title{Creates a form schema from a data set by guessing the field types required}
\usage{
createFormSchemaFromData(
  x,
  databaseId,
  label,
  folderId = databaseId,
  keyColumns = character(),
  requiredColumns = keyColumns,
  logicalAsSingleSelect = TRUE,
  logicalText = c("True", "False"),
  upload = FALSE
)
}
\arguments{
\item{x}{the data.frame or tibble for which to create form fields and a form
schema}

\item{databaseId}{the id of the database to which the form should belong.}

\item{label}{the label of the new form}

\item{folderId}{the id of the folder where the form should reside; defaults to the database id}

\item{keyColumns}{a character vector of the column names of the form fields that should be form keys}

\item{requiredColumns}{a character vector of the column names of the form fields that should be required}

\item{logicalAsSingleSelect}{by default TRUE and converts logical columns in the data frame to a single select form field; if FALSE then it will convert TRUE to 1 and FALSE to 0}

\item{logicalText}{the single select replacement values for c(TRUE, FALSE); default is c("True","False")}

\item{upload}{immediately upload the new form}
}
\description{
This function helps to create new form schemas for existing datasets.
}
