% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{deleteDatabaseUser}
\alias{deleteDatabaseUser}
\title{deleteDatabaseUser}
\usage{
deleteDatabaseUser(databaseId, userId)
}
\arguments{
\item{databaseId}{the id of the database}

\item{userId}{the (numeric) id of the user to remove from the database.}
}
\description{
Deletes a user from a database.
}
\details{
The user will receive a notification that their permission to access
the database has been revoked.
}
\examples{
\dontrun{

# Get the list of users in the database
databaseId <- "cxy12345gh"
users <- getDatabaseUsers(databaseId)

# Find the user with the email "bob@example.com"
matching <- sapply(users, function(u) u$email == "bob@example.com")
bob <- users[[which(matching)]]

# Remove the user from the database
deleteDatabaseUser(databaseId = databaseId, userId = bob$userId)

# You could also remove all users
for(user in users) {
  deleteDatabaseUser(databaseId = databaseId, userId = user$userId)
}
}

}
