% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{deleteForm}
\alias{deleteForm}
\title{Delete a form}
\usage{
deleteForm(databaseId, formId)
}
\arguments{
\item{databaseId}{The id of the database containing the form}

\item{formId}{The id of the form}
}
\description{
Deletes a form given the form id and the database id of the
database containing that form.
}
