% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{deleteFormField}
\alias{deleteFormField}
\alias{deleteFormField.character}
\alias{deleteFormField.formSchema}
\title{Delete a form field}
\usage{
deleteFormField(...)

\method{deleteFormField}{character}(formId, id, code, label, upload = FALSE, ...)

\method{deleteFormField}{formSchema}(formSchema, id, code, label, upload = FALSE, ...)
}
\arguments{
\item{...}{ignored}

\item{formId}{The identifier of the form online (provide either a formId or formSchema)}

\item{id}{The id of the form field (provide either an id, code, or label)}

\item{code}{The code of the form field (provide either an id, code, or label)}

\item{label}{The label of the form schema (provide either an id, code, or label)}

\item{upload}{Default is FALSE. If TRUE the modified form schema will be uploaded.}

\item{formSchema}{The offline schema of the form (provide either a formId or formSchema)}
}
\value{
The form field schema after the deletion. This will be the form field schema from the server if changes are uploaded.
}
\description{
Deletes a form field in an offline form schema or else downloads the form 
schema and deletes form field. Note that the either the upload argument 
must be TRUE for the field to be automatically deleted online. Otherwise, use 
updateFormSchema() to upload the changes after they are completed.
}
