% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{deleteRecord}
\alias{deleteRecord}
\title{Delete a record}
\usage{
deleteRecord(formId, recordId)
}
\arguments{
\item{formId}{a form id}

\item{recordId}{a record id}
}
\description{
This call deletes a single record for the given \code{formId} and
\code{recordId}.
}
