% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{formFieldSchema}
\alias{formFieldSchema}
\alias{textFieldSchema}
\alias{barcodeFieldSchema}
\alias{serialNumberFieldSchema}
\alias{quantityFieldSchema}
\alias{multilineFieldSchema}
\alias{dateFieldSchema}
\alias{weekFieldSchema}
\alias{fortnightFieldSchema}
\alias{monthFieldSchema}
\alias{singleSelectFieldSchema}
\alias{multipleSelectFieldSchema}
\alias{attachmentsFieldSchema}
\alias{calculatedFieldSchema}
\alias{subformFieldSchema}
\alias{referenceFieldSchema}
\alias{geopointFieldSchema}
\alias{userFieldSchema}
\alias{sectionFieldSchema}
\title{Create a simple form field schema}
\usage{
formFieldSchema(
  type,
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE,
  typeParameters = NULL
)

textFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

barcodeFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

serialNumberFieldSchema(
  label,
  description = NULL,
  digits = 5,
  prefixFormula = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

quantityFieldSchema(
  label,
  description = NULL,
  units = "",
  aggregation = "SUM",
  code = NULL,
  id = cuid(),
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

multilineFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

dateFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

weekFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

fortnightFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

monthFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

singleSelectFieldSchema(
  label,
  description = NULL,
  options = list(),
  presentation = "automatic",
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

multipleSelectFieldSchema(
  label,
  description = NULL,
  options = list(),
  presentation = "automatic",
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

attachmentsFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

calculatedFieldSchema(
  label,
  description = NULL,
  formula,
  code = NULL,
  id = cuid(),
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

subformFieldSchema(
  label,
  description = NULL,
  subformId,
  code = NULL,
  id = cuid(),
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

referenceFieldSchema(
  label,
  description = NULL,
  formId,
  code = NULL,
  id = cuid(),
  key = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

geopointFieldSchema(
  label,
  description = NULL,
  requiredAccuracy = NULL,
  manualEntryAllowed = TRUE,
  code = NULL,
  id = cuid(),
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

userFieldSchema(
  label,
  description = NULL,
  databaseId,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)

sectionFieldSchema(label, description = NULL)
}
\arguments{
\item{type}{The type character string defining the form field type.}

\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{key}{Whether the form field is a key field; default is FALSE}

\item{required}{Whether the form field is required; default is FALSE}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRules}{Relevance rules for the form field given as a character string; default is ""}

\item{validationRules}{Validation rules for the form field given as a character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}

\item{typeParameters}{The type parameters object specific to the type given.}

\item{digits}{The number of digits in the serial number}

\item{prefixFormula}{A formula as a character string defining the prefix for 
the serial number}

\item{units}{A character string describing the units, e.g. "litres per day"}

\item{aggregation}{A character string giving the aggregation function; "SUM" 
is default}

\item{options}{A list of the multiple select field options}

\item{presentation}{Default is "automatic"}

\item{formula}{A character string with the calculation formula}

\item{subformId}{The id of the sub-form}

\item{formId}{The id of the referenced form}

\item{requiredAccuracy}{Minimum required accuracy in metres/meters.}

\item{manualEntryAllowed}{Whether coordinates may be added manually; default
is TRUE}

\item{databaseId}{The database id of the form and users}
}
\description{
This is the function to create a basic offline form field schema. It is 
recommended to use the specific functions for each schema type such as 
textFieldSchema() or serialNumberFieldSchema().

You can define the format of the text that the users should type in a Text 
field using an Input Mask. 
See: https://www.activityinfo.org/support/docs/m/93526/l/1143998.html

Only one serial number field is possible in a form. The Prefix Formula is 
available for Serial Number fields and can be used to customise how the 
Serial Number will appear. In Subforms, the Prefix Formula can also derive 
from the Parent Form.

A Quantity field allow users to enter a numerical value. You can define the 
units and the aggregation function.

Multi-Line Text fields can be used to collect long answers to open-ended 
questions. They could be used for example to collect Comments about a 
specific Form or an Extended Narrative. It cannot be a key field.

The Date format in ActivityInfo is YYYY-MM-DD so no matter the way the Date 
is typed by a user it will always appear in this format.

The Week format in ActivityInfo is YYYY-WW. Users can directly type using 
this format or use the calendar to select a week. Please note that the Week 
field uses the EPI week convention.

The Fortnight field format in ActivityInfo is YYYYWW-WW (e.g. 2020W3-W4). 
Users can directly type using this format or use the calendar to select a 
week.

The Month format in ActivityInfo is YYYY-MM.

There is an options parameter for the list of single select items. Single 
Selection fields can be used to ask from users to select one out of two or 
more options as an answer. This can be a "Yes" or "No" question for example 
or any other type of question that can be answered with one option. Single 
Selection fields can be used in Prefix Formulas to customize Serial Numbers 
too.

There is an options parameter for the list of multiple select items. Multiple 
Selection fields can be used to ask from users to select any number of 
options as an answer.

An attachments field allow users to add one or more attachments.

A calculated field can be used to define a formula that presents calculation 
results in the form.

A subform field can be used to define a field that contains a subform.

A reference field can be used to make reference to a record in another form.

A Geographic Point field allow users to enter a geo-location with a certain
accuracy.

User fields allow you to select a specific user from a list. It is a field 
useful for the Case Management Database template with which you can assign 
Cases to specific users. User fields and inactive users: If you are using 
the User field in your Form and a user who has added Records to the Form, is 
no longer active, instead of deleting that user, create a new Role empty of 
permissions and assign it to that user.

A special form field to define a section header for the form.
}
\details{
A quantity field cannot be a key field.

An attachments field cannot be a key field.

A calculated field cannot be a key field.

A subform field cannot be a key field.

The user interface as the required geo-location accuracy options:
* No minimum accuracy
* Medium accuracy (25 m); default
* High accuracy (10 m)"

This field cannot be a key field.
}
