% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{formSchema}
\alias{formSchema}
\title{Create a form schema object}
\usage{
formSchema(
  databaseId,
  label,
  id = cuid(),
  elements = list(),
  parentFormId = NULL,
  folderId = NULL
)
}
\arguments{
\item{databaseId}{The identifier of the database containing the form}

\item{label}{The label of the form}

\item{id}{The identifier of the form; if unused will generate a new cuid}

\item{elements}{The elements/form fields of the form}

\item{parentFormId}{The identifier of the parent form of the new sub-form.}

\item{folderId}{The identifier of the folder containing the new form.}
}
\description{
Generates a new form schema object which can be used to add a new form to
ActivityInfo with an optional folder specified or in the case of a  new sub-
form with a parent form specified.
}
\examples{
survey <- formSchema(
   databaseId = "cyx12345gh",
   label = "ActivityInfo form generated from R",
   elements = list(
     textFieldSchema(label = "What is your name?", code = "name", required = TRUE),
     dateFieldSchema(label = "When were you born?", code = "dob")))
\dontrun{
addForm(survey)
}

}
