% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{fromActivityInfoJson}
\alias{fromActivityInfoJson}
\title{Wraps jsonlite::fromJSON to match the style of JSON produced
by the ActivityInfo server.}
\usage{
fromActivityInfoJson(x)
}
\arguments{
\item{x}{a json string or a httr response object to convert to an R object}
}
\description{
Wraps jsonlite::fromJSON to match the style of JSON produced
by the ActivityInfo server.
}
