% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{getAdminLevelEntities}
\alias{getAdminLevelEntities}
\title{Get all entities belonging to an administrative level by ID}
\usage{
getAdminLevelEntities(id)
}
\description{
Retrieves all entities belonging to an administrative level with the given
identifier.
}

