% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getAttachment}
\alias{getAttachment}
\title{Gets an attachment}
\usage{
getAttachment(formId, recordId, fieldId, blobId, extension)
}
\arguments{
\item{formId}{a form id}

\item{recordId}{the record Id}

\item{fieldId}{the attachment field id}

\item{blobId}{the attachment blob id}
}
\description{
Retrieve an attachment and store it to a temporary file.
}
