% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{getAuthorizedUsers}
\alias{getAuthorizedUsers}
\title{Retrieves a list of authorized users of the given database}
\usage{
getAuthorizedUsers(databaseId)
}
\arguments{
\item{databaseId}{the id of the database for which to retrieve authorized users}
}
\description{
Retrieves a list of authorized users of the given database
}

