% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingAccount}
\alias{getBillingAccount}
\title{getBillingAccount}
\usage{
getBillingAccount(billingAccountId, asDataFrame = TRUE)
}
\arguments{
\item{billingAccountId}{Billing ID}

\item{asDataFrame}{Output as data.frame, Default: TRUE}
}
\value{
Billing account information in list or data.frame output
}
\description{
Get billing account information
}
