% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingAccountDatabases}
\alias{getBillingAccountDatabases}
\title{getBillingAccountDatabases}
\usage{
getBillingAccountDatabases(asDataFrame = TRUE, billingId)
}
\arguments{
\item{asDataFrame}{Output as data.frame, Default: TRUE}

\item{billingId}{Billing ID}
}
\value{
Information on databases under billing account in list or data.frame output
}
\description{
Data for all databases under billing account
}
