% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingAccountDatabases}
\alias{getBillingAccountDatabases}
\title{getBillingAccountDatabases}
\usage{
getBillingAccountDatabases(billingAccountId, asDataFrame = TRUE)
}
\arguments{
\item{billingAccountId}{Billing ID}

\item{asDataFrame}{Output as data.frame, Default: TRUE}
}
\value{
Information on databases under billing account in list or data.frame output
}
\description{
Data for all databases under billing account
}
