% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getBillingAccountUsers}
\alias{getBillingAccountUsers}
\title{getBillingAccountUsers}
\usage{
getBillingAccountUsers(asDataFrame = TRUE, billingId)
}
\arguments{
\item{asDataFrame}{Output as data.frame, Default: TRUE}

\item{billingId}{Billing ID}
}
\value{
Billing account user(s) in list or data.frame output
}
\description{
Billing account users
}
