% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingInfo.R
\name{getDatabaseBillingAccount}
\alias{getDatabaseBillingAccount}
\title{getDatabaseBillingAccount}
\usage{
getDatabaseBillingAccount(databaseId, asDataFrame = TRUE)
}
\arguments{
\item{databaseId}{Database ID}

\item{asDataFrame}{Data.frame output, Default: TRUE}
}
\value{
Database owner in list output
}
\description{
Get database owner for a given database. This gives the owning organization rather than the user who created the database.
}
