% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{getDatabaseRecordTable}
\alias{getDatabaseRecordTable}
\title{Extract all Form Record Tables for a Database in "long" format.}
\usage{
getDatabaseRecordTable(database.id = NA, as.single.table = FALSE)
}
\arguments{
\item{database.id}{the numeric id of the database}

\item{as.single.table}{specify whether to merge all Form Record Tables into a single table}
}
\description{
Each Form Record Table is returned as a data.frame, with:
-> Columns for all Form Dimensions, including:
   - Sub-Form Name, Sub-Form Id, Sub-Form Record Id and Sub-Form Period (where applicable)
   - Reference Field Dimensions (where applicable)
-> Rows for every measured value, with the measure field name under "Measure" and the measured value under "Value"
}
\details{
By default, returns a list of all Form Record Tables.
If @param as.single.table is TRUE, all Form Record Tables will be concatenated into a single table, with missing values given as <NA>.
}
