% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabases}
\alias{getDatabases}
\title{getDatabases()}
\usage{
getDatabases()
}
\description{
Retrieves a list of databases the authenticated user owns, or has been shared
with the authenticated user as a
}
