% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabases}
\alias{getDatabases}
\title{getDatabases()}
\usage{
getDatabases(asDf = FALSE)
}
\arguments{
\item{asDf}{Default is false but if TRUE will return the list of databases as a dataframe.}
}
\description{
Retrieves a list of databases the authenticated user owns, or has been shared
 with
}
