% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseSchema}
\alias{getDatabaseSchema}
\title{getDatabaseSchema}
\usage{
getDatabaseSchema(databaseId)
}
\arguments{
\item{databaseId}{database identifier}
}
\description{
This function is deprecated in favor of getDatabaseTree(). Please use \link{getDatabaseTree}.
}
\examples{
\dontrun{
getDatabaseSchema("ck2k93muu2")
}
}
