% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDatabaseSchema}
\alias{getDatabaseSchema}
\title{GetDatabaseSchema}
\usage{
getDatabaseSchema(databaseId)
}
\arguments{
\item{databaseId}{database identifier}
}
\description{
Retrieves the schema (partners, activities, indicators and attributes) from
for the given database
}

