% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{getDatabaseUsers}
\alias{getDatabaseUsers}
\title{getDatabaseUsers}
\usage{
getDatabaseUsers(databaseId, asDataFrame = TRUE)
}
\arguments{
\item{databaseId}{The database ID}

\item{asDataFrame}{Retrieve user list as a data.frame, otherwise returns as list. Default: TRUE}
}
\description{
Retrieves the list of users with access to the database in a data.frame format.
}
