% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{getDatabaseValueTable}
\alias{getDatabaseValueTable}
\title{Extract a full database in "long" format}
\usage{
getDatabaseValueTable(database.id = NA, include.comments = FALSE,
  col.names = NULL)
}
\arguments{
\item{database.id}{the numeric id of the database}

\item{include.comments}{TRUE if narrative comments fields should also be included}

\item{col.names}{a named character vector containing alternate names for the resulting table}
}
\description{
The resulting data.frame includes a row for each reported indicator
and month, with columns for the site's location, the administrative levels,
and for each of the single-valued attributes defined.
}

