% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{getFormSchema}
\alias{getFormSchema}
\title{Queries the schema of a form or extracts the schema from the remote records 
table}
\usage{
getFormSchema(formId)
}
\arguments{
\item{formId}{the form identifier}
}
\value{
A list with class \sQuote{formSchema}.
}
\description{
The result has a class "formSchema" and can be transformed to
data.frame using `as.data.frame()`
}
\examples{
\dontrun{
formSchema <- getFormSchema("ck2lt9wp3g")
formSchemaTable <- as.data.frame(getFormSchema("ck2lt9wp3g"))
}
}
