% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{getFormSchema}
\alias{getFormSchema}
\title{Queries the schema of a form}
\usage{
getFormSchema(formId)
}
\arguments{
\item{formId}{the form identifier}
}
\value{
A list with class \sQuote{formSchema}.
}
\description{
Queries the schema of a form
}
\examples{
\dontrun{
getFormSchema("ck2lt9wp3g")
}
}
