% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLong.R
\name{getQuantityTable}
\alias{getQuantityTable}
\title{Extract all quantities from a database in \dQuote{long} format}
\usage{
getQuantityTable(databaseId = NA, folderId)
}
\arguments{
\item{database.id}{the CUID of the database (e.g., "ck2yrizmo2" or "d00000006583")}

\item{folder.id}{(optional) the id of the folder or form to include. If omitted, the whole database will be included in the export.}
}
\value{
a single data.frame with quantity values in rows, and dimensions in columns.
}
\description{
Extract all quantities from a database in \dQuote{long} format
}
