% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getRecordHistory}
\alias{getRecordHistory}
\title{Gets the list of changes to a record}
\usage{
getRecordHistory(formId, recordId)
}
\arguments{
\item{formId}{a form id}

\item{recordId}{a record id}
}
\description{
This calls retrieves a list of all changes to the record, and the users
who are
}
