% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getRecords}
\alias{getRecords}
\title{Get a table of records}
\usage{
getRecords()
}
\arguments{
\item{formId}{a form id}

\item{recordId}{a record id}
}
\description{
This function will create a reference to records on the server. You can use this like a data.frame or to download the table use the collect() function.
}
