% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getRecords}
\alias{getRecords}
\title{Get a form's records as a table}
\usage{
getRecords(form, style)
}
\arguments{
\item{form}{a form id, form schema, form tree, or activity info data frame}

\item{style}{a column style object that defines how table columns should be created from a form; use \link{columnStyle} to create a new style;
default column styles can be set with an option or \link{defaultColumnStyle}. Default \code{columnNames} is set to "pretty".}
}
\description{
This function will create a reference to records on the server. For large forms,
this does not immediately transfer all records; you can use the \link[dplyr:filter]{dplyr::filter},
\link[dplyr:arrange]{dplyr::arrange}, \link[dplyr:slice]{dplyr::slice_head}, \link[dplyr:slice]{dplyr::slice_tail}, and/or \link{adjustWindow}
functions to narrow the the selection of records.

To prevent downloading the records before making a selection use the verbs in
this order.
\enumerate{
\item \link[dplyr:arrange]{dplyr::arrange} (limited to a single column) and/or \link[dplyr:filter]{dplyr::filter} in any combination
\item \link[dplyr:slice]{dplyr::slice_head} or \link{adjustWindow} in any combination
}
}
\examples{
\dontrun{

# Retrieve all the records from the Simple 3W template's project form:
# https://www.activityinfo.org/app#form/ceam1x8kq6ikcujg/table

records <- getRecords("ceam1x8kq6ikcujg")

# Now, filter by only projects with a status "Under implementation"
records <- getRecords("ceam1x8kq6ikcujg") \%>\%
   filter(`Project status` == "Under implementation")

# If you want to use the field codes, set the column style using the `columnStyle()`
# function and setting `columnName` argument to 'code'.
# You can then use the field code in your filter:
records <- getRecords("ceam1x8kq6ikcujg", style = columnStyle(columnNames = "code")) \%>\%
   filter(START_MONTH == "2018-01")

# If you prefer to work with the ActivityInfo field IDs, then you can change
# the column style and the `columnNames` argument to "id":
records <- getRecords("ceam1x8kq6ikcujg", style = columnStyle(columnNames = "id")) \%>\%
   filter(cgkh1k5kq6k0gsmv == "2018-01")

}

}
