% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{idColumnStyle}
\alias{idColumnStyle}
\title{A form table style including all columns with ids as column names}
\usage{
idColumnStyle()
}
\description{
This function is primarily used in \link{getRecords}.
}
