% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTable}
\alias{importTable}
\title{Deprecated function to batch import a data.frame into an ActivityInfo form; use \link{importRecords}.}
\usage{
importTable(...)
}
\arguments{
\item{...}{parameters of importRecords()}
}
\description{
Deprecated function to batch import a data.frame into an ActivityInfo form; use \link{importRecords}.
}
