% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTable}
\alias{importTable}
\title{Batch imports a data.frame into an ActivityInfo form}
\usage{
importTable(formId, data, recordIdColumn, parentIdColumn)
}
\arguments{
\item{formId}{The id of the form to which the records should be imported}

\item{data}{the data.frame to import}

\item{recordIdColumn}{optional. If provided, this column will be used to provide
the ActivityInfo record ids.}
}
\description{
Batch imports a data.frame into an ActivityInfo form
}
