% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTable}
\alias{importTable}
\title{Batch imports a data.frame into an ActivityInfo form}
\usage{
importTable(formId, data, recordIdColumn, parentIdColumn)
}
\arguments{
\item{formId}{The form ID}

\item{data}{The data.frame to import}

\item{recordIdColumn}{The record ID column}

\item{parentIdColumn}{The parent ID column required when importing a subform}
}
\description{
Batch imports a data.frame into an ActivityInfo form
}
