% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTable}
\alias{importTable}
\title{Batch imports a data.frame into an ActivityInfo form}
\usage{
importTable(formId, data, recordIdColumn, parentIdColumn)
}
\description{
Batch imports a data.frame into an ActivityInfo form
}
