\name{lookupCountryId}
\alias{lookupCountryId}
\title{Looks up the internal identifier for a country by code}
\usage{
lookupCountryId(countryId)
}
\arguments{
  \item{countryId}{the country's name or ISO-3166-2 code}
}
\value{
the country's internal identifier used by AI.
}
\description{
Looks up the internal identifier for a country by code
}

