% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{minimalColumnStyle}
\alias{minimalColumnStyle}
\title{A form table style with columns limited to those in the web interface}
\usage{
minimalColumnStyle(maxDepth = 2)
}
\arguments{
\item{maxDepth}{the maximum depth of recursion in referenced forms. This is set to 3 by default.}
}
\description{
This function is primarily used in getRecords().
}
