% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.r
\name{multilineFieldSchema}
\alias{multilineFieldSchema}
\title{Create a multi-line or narrative form field schema}
\usage{
multilineFieldSchema(
  label,
  description = NULL,
  code = NULL,
  id = cuid(),
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{required}{Whether the form field is required; default is FALSE}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRules}{Relevance rules for the form field given as a character string; default is ""}

\item{validationRules}{Validation rules for the form field given as a character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
Multi-Line Text fields can be used to collect long answers to open-ended 
questions. They could be used for example to collect Comments about a 
specific Form or an Extended Narrative. It cannot be a key field.
}
