% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{namedElementVarList}
\alias{namedElementVarList}
\title{Get form field schemas from a form tree in a named list using a column style 
to select and name each form field}
\usage{
namedElementVarList(formTree, style = defaultColumnStyle())
}
\arguments{
\item{style}{a column style object.}

\item{x}{the form tree object.}
}
\description{
This helper function provides form field schemas in a named list. This can be 
useful for examining and manipulating form fields. See also copySchema().
}
