% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{quantityFieldSchema}
\alias{quantityFieldSchema}
\title{Create a quantity form field schema}
\usage{
quantityFieldSchema(
  label,
  description = NULL,
  units = "",
  aggregation = "SUM",
  code = NULL,
  id = cuid(),
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRule = "",
  validationRule = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{units}{A character string describing the units, e.g. "litres per day"}

\item{aggregation}{A character string giving the aggregation function; "SUM" 
is default}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{required}{Whether the form field is required; default is FALSE}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRule}{Relevance rules for the form field given as a single character string; default is ""}

\item{validationRule}{Validation rules for the form field given as a single character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
A Quantity field allow users to enter a numerical value. You can define the 
units and the aggregation function.
}
\details{
A quantity field cannot be a key field.
}
