% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{queryAuditLog}
\alias{queryAuditLog}
\title{Query the audit log of a database}
\usage{
queryAuditLog(
  databaseId,
  before = Sys.time(),
  after,
  resourceId = NULL,
  typeFilter = NULL,
  limit = 1000
)
}
\arguments{
\item{databaseId}{the database identifier.}

\item{before}{a \emph{Date} or \emph{POSIX} to filter result before a given
time; defaults to the time of the query.}

\item{after}{an optional \emph{Date} or \emph{POSIX}}

\item{typeFilter}{a character string with the event type to filter on; default is none.}

\item{limit}{the maximum number of events to return. Default is 1,000 events.}

\item{resoureId}{a resource (i.e. form or folder) identifier to filter on.}
}
\value{
A data frame with the results of the query and with three attributes as described in the \emph{Details}
section.
}
\description{
Query the audit log of a database
}
\details{
A database audit log contains details on a wide variety of events in the database. These events include the
deletion of records or adding a user. This function queries the audit log and returns the event details as a data
frame. Note that each query returns a maximum of 100 events, therefore the data frame has three attributes to help
you query older log entries:
\itemize{
\item \emph{databaseId}: the database identifier;
\item \emph{moreEvents}: if \code{TRUE} then there are more (older) entries to be queried;
\item \emph{startTime}: the (UNIX) timestamp which is essentially the same as the time provided in the \code{before}
argument of the query;
\item \emph{endTime} the earliest timestamp of the range, which can be used to query the next set of entries.
}

The following event type filters are supported:
\itemize{
\item \emph{RECORD}: changes to records
\item \emph{FORM}: changes to form schemas
\item \emph{FOLDER}: changes to folders
\item \emph{DATABASE}: changes to the database itself (renaming, creation)
\item \emph{LOCK}: changes to the database locks
\item \emph{USER_PERMISSION}: changes to user permissions
\item \emph{ROLE}: changes to roles
}
}
\examples{
\dontrun{

# Query up to 10,000 records
events <- queryAuditLog("cax132253", limit = 10000)

# Query events after the beginning of 2022
events <- queryAuditLog("cax132253", after = as.Date("2022-01-01"))

# Query only changes to user permissions
events <- queryAuditLog("cax132253", typeFilter = "USER_PERMISSION")
}
}
