% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableQuery.R
\name{queryTable}
\alias{queryTable}
\title{Queries a form as a flat, two-dimensional table}
\usage{
queryTable(form, columns, ..., truncate.strings = TRUE)
}
\arguments{
\item{columns}{select columns}

\item{truncate.strings}{TRUE if longer strings should be truncated to 128 characters}

\item{the}{form to query. This can be an object of type "tree", "class", or the id of the
form as a character.}
}
\description{
Queries a form as a flat, two-dimensional table
}
\examples{
\dontrun{
queryTable("a2145507918", columns = c(
id="_id",
name="Name",
campId="camp",
camp="Camp.Name",
governorate="Camp.Governorate.Name",
teachers="teachers",
students="students",
type="[Type of School]"
))
}
}
\references{
\href{ActivityInfo Formulas Manual}{http://help.activityinfo.org/m/77022}
}
