% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableQuery.R
\name{queryTable}
\alias{queryTable}
\title{Queries a form as a flat, two-dimensional table.}
\usage{
queryTable(form, columns, ...)
}
\arguments{
\item{the}{form to query. This can be an object of type "tree", "class", or the id of the 
form as a character.}
}
\description{
Queries a form as a flat, two-dimensional table.
}
