% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableQuery.R
\name{queryTable}
\alias{queryTable}
\title{Queries a form as a flat, two-dimensional table}
\usage{
queryTable(
  form,
  columns,
  ...,
  truncateStrings = TRUE,
  asTibble = FALSE,
  makeNames = TRUE,
  asUI = FALSE,
  truncate.strings = truncateStrings,
  filter,
  window
)
}
\arguments{
\item{form}{form to query. This can be an object of type "tree", "class", or the id of the
form as a character.}

\item{columns}{select columns, see Details}

\item{...}{If columns parameter is empty, the additional arguments are used as columns.}

\item{truncateStrings}{TRUE if longer strings should be truncated to 128 characters}

\item{truncate.strings}{Deprecated: please use truncateStrings. TRUE if longer strings should be truncated to 128 characters}

\item{filter}{an ActivityInfo formula string that limits the records returned}
}
\description{
Queries a form as a flat, two-dimensional table
}
\details{
To select columns, you can use
\itemize{
  \item \code{_id} to get the record identifier,
  \item \code{_lastEditTime} to get the (Unix) timestamp (in seconds),
  \item the identifier of the form field (e.g. \code{ck849fbaw4}),
  \item the field code (e.g. \code{TOT_BNF}),
  \item the field label, enclosed in brackets if it contains spaces (e.g. \code{[Type of school]})
}
If you want to select a column in a reference form or in a parent form, use a period to create a path to the desired
column. For example:
\itemize{
  \item \code{parent._id} will select the identifier of the parent record and
  \item \code{Camp.Name} will select the \sQuote{Name} field in the \sQuote{Camp} form.
}
}
\examples{
\dontrun{
queryTable("a2145507918", columns = c(
  id = "_id",
  name = "Name",
  campId = "camp",
  camp = "Camp.Name",
  governorate = "Camp.Governorate.Name",
  teachers = "teachers",
  students = "students",
  type = "[Type of School]"
))
}
}
\references{
Unix time, Wikipedia \url{https://en.wikipedia.org/wiki/Unix_time}
}
