% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{recoverRecord}
\alias{recoverRecord}
\title{Recover a deleted record}
\usage{
recoverRecord(formId, recordId)
}
\arguments{
\item{formId}{a form id}

\item{recordId}{a record id}
}
\description{
This call recovers a single deleted record for the given \code{formId} and
\code{recordId}.
}
