% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{roleAssignment}
\alias{roleAssignment}
\title{roleAssignment}
\usage{
roleAssignment(roleId, roleParameters = list(), roleResources)
}
\arguments{
\item{roleId}{the id of the role to assign to the user}

\item{roleParameters}{a named list of parameters, if the role has any parameters}

\item{roleResources}{the list of resources (database, folder, form, or report)
to assign to this user. Using the databaseId assigns all resources to this user}
}
\description{
Creates a role assignment object
}
\examples{
{
  # Role assignment for a reporting role with a partner parameter
  roleAssignment(
    roleId = "rp",
    roleParameters = list(partner = reference(formId = "cxadsfs32", recordId = "c3423423")),
    roleResources = "cxa99335"
  )


  # Role assignment for an administrator role without any role parameters
  roleAssignment(
    roleId = "admin",
    roleResources = c("cxa99335", "c8234234")
  )
}

}
