% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{sectionFieldSchema}
\alias{sectionFieldSchema}
\title{Create a section header form field schema}
\usage{
sectionFieldSchema(label, description = NULL)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}
}
\description{
A special form field to define a section header for the form.
}
\seealso{
Other field schemas: 
\code{\link{attachmentFieldSchema}()},
\code{\link{barcodeFieldSchema}()},
\code{\link{calculatedFieldSchema}()},
\code{\link{dateFieldSchema}()},
\code{\link{formFieldSchema}()},
\code{\link{geopointFieldSchema}()},
\code{\link{monthFieldSchema}()},
\code{\link{multilineFieldSchema}()},
\code{\link{multipleSelectFieldSchema}()},
\code{\link{quantityFieldSchema}()},
\code{\link{referenceFieldSchema}()},
\code{\link{serialNumberFieldSchema}()},
\code{\link{singleSelectFieldSchema}()},
\code{\link{subformFieldSchema}()},
\code{\link{userFieldSchema}()},
\code{\link{weekFieldSchema}()}
}
\concept{field schemas}
