% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{serialNumberFieldSchema}
\alias{serialNumberFieldSchema}
\title{Create a serial number form field schema}
\usage{
serialNumberFieldSchema(
  label,
  description = NULL,
  digits = 5,
  prefixFormula = NULL,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{digits}{The number of digits in the serial number}

\item{prefixFormula}{A formula as a character string defining the prefix for 
the serial number}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{key}{Whether the form field is a key field; default is FALSE}

\item{required}{Whether the form field is required; default is FALSE}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRules}{Relevance rules for the form field given as a character string; default is ""}

\item{validationRules}{Validation rules for the form field given as a character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
Only one serial number field is possible in a form. The Prefix Formula is 
available for Serial Number fields and can be used to customise how the 
Serial Number will appear. In Subforms, the Prefix Formula can also derive 
from the Parent Form.
}
