% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.r
\name{serialNumberFieldSchema}
\alias{serialNumberFieldSchema}
\title{Create a text form field schema}
\usage{
serialNumberFieldSchema(..., digits = 5, prefixFormula = NULL)
}
\arguments{
\item{...}{See arguments defined in \link[activityinfo]{formFieldSchema}}

\item{digits}{The number of digits in the serial number}

\item{prefixFormula}{A formula as a character string defining the prefix for 
the serial number}
}
\description{
Create a text form field schema
}
