% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.r
\name{singleSelectFieldSchema}
\alias{singleSelectFieldSchema}
\title{Create a Single Select form field schema}
\usage{
singleSelectFieldSchema(
  label,
  description = NULL,
  values = list(),
  presentation = "automatic",
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = FALSE,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRules = "",
  validationRules = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{values}{A list of the single select field values}

\item{presentation}{Default is "automatic"}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{key}{Whether the form field is a key field; default is FALSE}

\item{required}{Whether the form field is required; default is FALSE}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRules}{Relevance rules for the form field given as a character string; default is ""}

\item{validationRules}{Validation rules for the form field given as a character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
There is an options parameter for the list of single select items. Single 
Selection fields can be used to ask from users to select one out of two or 
more options as an answer. This can be a "Yes" or "No" question for example 
or any other type of question that can be answered with one option. Single 
Selection fields can be used in Prefix Formulas to customize Serial Numbers 
too.
}
