% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{stageImport}
\alias{stageImport}
\title{Stages data to import to ActivityInfo}
\usage{
stageImport(text, direct = TRUE)
}
\arguments{
\item{text}{The text of the file to import.}

\item{direct}{Whether the import should be directly staged to Google Cloud Storage. This may not be possible if connecting from Syria or other countries that are blocked from accessing Google services directly. This option is ignored when connecting to a self-managed instance of ActivityInfo.}
}
\description{
Stages data to import to ActivityInfo
}
