% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.r
\name{textFieldSchema}
\alias{textFieldSchema}
\title{Create a text form field schema}
\usage{
textFieldSchema(...)
}
\arguments{
\item{...}{See arguments defined in \link[activityinfo]{formFieldSchema}}
}
\description{
Create a text form field schema
}
