% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{toActivityInfoFormula}
\alias{toActivityInfoFormula}
\title{Convert an expression using columns in a remote records into an ActivityInfo style formula}
\usage{
toActivityInfoFormula(.data, expr)
}
\arguments{
\item{.data}{the remote records object fetched with getRecords().}

\item{expr}{the expression to convert}
}
\description{
This function attempts to convert an R expression using the columns of the getRecords() object into an ActivityInfo formula (as a string) using the form field ids for use in queryTable() or in the web interface.
}
