% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{toSelectOptions}
\alias{toSelectOptions}
\title{Converts a factor, list or vector to select options}
\usage{
toSelectOptions(options)
}
\arguments{
\item{options}{The character vector or list to convert to options}
}
\description{
Creates a list of options in the ActivityInfo format for Single Select or
Multiple Select form fields.
}
