% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{updateGrant}
\alias{updateGrant}
\title{updateGrant}
\usage{
updateGrant(databaseId, userId, resourceId, permissions)
}
\arguments{
\item{databaseId}{the id of the database}

\item{userId}{the (numeric) id of the user to update}

\item{resourceId}{the id of the form or folder}

\item{permissions}{the permissions to grant to the user for the given resource}
}
\description{
Adds or updates a grant for a user to a specific resource.
See \link{resourcePermissions} for how to set resource-level permissions for
a grant.
}
\examples{
\dontrun{
updateGrant(
  databaseId = "cxy123", user = 165,
  resourcePermissions(
    add_record = TRUE,
    edit_record = TRUE,
    delete_record = TRUE
  )
)
}

}
