% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{updateRole}
\alias{updateRole}
\title{Updates a role's definition in the database}
\usage{
updateRole(databaseId, role)
}
\arguments{
\item{databaseId}{the id of the database}

\item{role}{the role definition,}
}
\description{
Updates a role's definition in the database
}
\examples{
\dontrun{
updateRole("cxy123", list(
   id = "rp",
   label = "Reporting partner",
  permissions = list(
  list(
    operation = "VIEW",
    filter = "ck5dxt1712 == @user.partner"),
  list(
    operation = "EDIT_RECORD",
    filter = "ck5dxt1712 == @user.partner",
    securityCategories = list()
  ),
  list(
    operation = "EXPORT_RECORDS"
  )
),
parameters = list(
  list(
    parameterId = "partner",
    label = "Partner",
    range = "ck5dxt1712"
  )
),
filters = list(
  list(id = "partner",
       label = "partner is user's partner",
       filter = "ck5dxt1712 == @user.partner")
)
))

}

}
