% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{updateRole}
\alias{updateRole}
\title{Updates a role's definition in the database
See \link{permissions} for more details.}
\usage{
updateRole(databaseId, role)
}
\arguments{
\item{databaseId}{the id of the database}

\item{role}{the role definition}
}
\description{
Updates a role's definition in the database
See \link{permissions} for more details.
}
\examples{
\dontrun{
updateRole("cxy123", list(
  id = "rp",
  label = "Reporting partner",
  permissions = permissions(
    view = "ck5dxt1712 == @user.partner",
    edit_record = "ck5dxt1712 == @user.partner",
    export_records = TRUE
  ),
  parameters = list(
    list(
      parameterId = "partner",
      label = "Partner",
      range = "ck5dxt1712"
    )
  ),
  filters = list(
    list(
      id = "partner",
      label = "partner is user's partner",
      filter = "ck5dxt1712 == @user.partner"
    )
  )
))
}

}
