% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formField.R
\name{userFieldSchema}
\alias{userFieldSchema}
\title{Create a User form field schema}
\usage{
userFieldSchema(
  label,
  description = NULL,
  databaseId,
  code = NULL,
  id = cuid(),
  key = FALSE,
  required = key,
  hideFromEntry = FALSE,
  hideInTable = FALSE,
  relevanceRule = "",
  validationRule = "",
  reviewerOnly = FALSE
)
}
\arguments{
\item{label}{The label of the form field}

\item{description}{The description of the form field}

\item{databaseId}{The database id of the form and users}

\item{code}{The code name of the form field}

\item{id}{The id of the form Field; default is to generate a new cuid}

\item{key}{Whether the form field is a key field; default is FALSE}

\item{required}{Whether the form field is required; default is FALSE}

\item{hideFromEntry}{Whether the form field is hidden during data entry; default is FALSE}

\item{hideInTable}{Whether the form field is hidden during data display; default is FALSE}

\item{relevanceRule}{Relevance rules for the form field given as a single character string; default is ""}

\item{validationRule}{Validation rules for the form field given as a single character string; default is ""}

\item{reviewerOnly}{Whether the form field is for reviewers only; default is FALSE}
}
\description{
User fields allow you to select a specific user from a list. It is a field
useful for the Case Management Database template with which you can assign
Cases to specific users. User fields and inactive users: If you are using
the User field in your Form and a user who has added Records to the Form, is
no longer active, instead of deleting that user, create a new Role empty of
permissions and assign it to that user.
}
\seealso{
Other field schemas: 
\code{\link{attachmentFieldSchema}()},
\code{\link{barcodeFieldSchema}()},
\code{\link{calculatedFieldSchema}()},
\code{\link{dateFieldSchema}()},
\code{\link{formFieldSchema}()},
\code{\link{geopointFieldSchema}()},
\code{\link{monthFieldSchema}()},
\code{\link{multilineFieldSchema}()},
\code{\link{multipleSelectFieldSchema}()},
\code{\link{quantityFieldSchema}()},
\code{\link{referenceFieldSchema}()},
\code{\link{sectionFieldSchema}()},
\code{\link{serialNumberFieldSchema}()},
\code{\link{singleSelectFieldSchema}()},
\code{\link{subformFieldSchema}()},
\code{\link{weekFieldSchema}()}
}
\concept{field schemas}
